+++ .github/workflows/darshan_ldms_test_ci.yml
@@ -51,6 +51,8 @@ jobs:
     - name: Install LDMS
       run: |
         cd ovis
+        git fetch --all
+        git checkout 084f77d
         sh autogen.sh
         set -e
         mkdir -p build
@@ -70,7 +72,7 @@ jobs:
         mkdir build
         cd build
         # build darshan against LDMS library
-        CC=mpicc ../configure --prefix=$DARSHAN_INSTALL_PATH --with-log-path-by-env=DARSHAN_LOGPATH --with-ldms=/opt/ovis-latest --with-jobid-env=NONE --enable-hdf5-mod
+        CC=mpicc ../configure --prefix=$DARSHAN_INSTALL_PATH --with-log-path-by-env=DARSHAN_LOGPATH --enable-ldms-mod --with-ldms=/opt/ovis-latest --with-jobid-env=NONE --enable-hdf5-mod
         make
         make install
     - name: Install darshan-util
