+++ darshan-runtime/configure.ac
@@ -422,16 +422,16 @@ if test "x$enable_darshan_runtime" = xyes ; then
    fi
    
    # LDMS support (disabled by default)
-   AC_ARG_WITH([ldms],
-      [AS_HELP_STRING([--with-ldms],
+   AC_ARG_ENABLE([ldms-mod],
+      [AS_HELP_STRING([--enable-ldms-mod],
                       [Include LDMS for runtme I/O monitoring])],
-      [], [with_ldms=no])
+      [AC_ARG_WITH(ldms, 
+	[AS_HELP_STRING([--with-ldms],
+                      [Installation directory for LDMS.])],[CHECK_LDMS], 
+	[AC_MSG_ERROR(m4_normalize([--enable-ldms-mod does not take any argument, use --with-ldms to
+                        provide the LDMS install prefix.]))])], [enable_ldms_mod=no])
 
-   AM_CONDITIONAL([HAVE_LDMS],[test "x$with_ldms" = xyes])
-   if test "x$with_ldms" != xno ; then
-	# check required LDMS libraries are available.
-	CHECK_LDMS
-   fi
+   AM_CONDITIONAL([HAVE_LDMS],[test "x$enable_ldms_mod" = xyes])
 
    # inform about PnetCDF installs not found in default locations
    AC_ARG_WITH([pnetcdf],
@@ -767,6 +767,7 @@ else
    enable_posix_mod=no
    enable_stdio_mod=no
    enable_dxt_mod=no
+   enable_ldms_mod=no
    enable_heatmap_mod=no
    enable_mpiio_mod=no
    enable_apmpi_mod=no
@@ -788,6 +789,7 @@ AM_CONDITIONAL(BUILD_NULL_MODULE,   [test "x$enable_null_mod"    = xyes])
 AM_CONDITIONAL(BUILD_POSIX_MODULE,  [test "x$enable_posix_mod"   = xyes])
 AM_CONDITIONAL(BUILD_STDIO_MODULE,  [test "x$enable_stdio_mod"   = xyes])
 AM_CONDITIONAL(BUILD_DXT_MODULE,    [test "x$enable_dxt_mod"     = xyes])
+AM_CONDITIONAL(BUILD_LDMS_MODULE,   [test "x$enable_ldms_mod"    = xyes])
 AM_CONDITIONAL(BUILD_MPIIO_MODULE,  [test "x$enable_mpiio_mod"   = xyes])
 AM_CONDITIONAL(BUILD_HDF5_MODULE,   [test "x$enable_hdf5_mod"    = xyes])
 AM_CONDITIONAL(BUILD_PNETCDF_MODULE,[test "x$enable_pnetcdf_mod" = xyes])
@@ -852,12 +854,7 @@ if test "x$enable_darshan_runtime" = xyes ; then
          enable_apmpi_mod="yes (collective sync mode)"
       fi
    fi
-   # add ldms details
-   if test "x$with_ldms" = xno ; then
-	enable_ldms="no"
-      else
-	enable_ldms="yes"
-   fi
+   
    echo "------------------------------------------------------------------------------
    ${PACKAGE_NAME} Version ${PACKAGE_VERSION} configured with the following features:"
    if test "x$ENABLE_MPI" = xyes ; then
@@ -882,7 +879,8 @@ if test "x$enable_darshan_runtime" = xyes ; then
            Lustre        module support  - $enable_lustre_mod
            MDHIM         module support  - $enable_mdhim_mod
            HEATMAP       module support  - $enable_heatmap_mod
-	   LDMS  Runtime I/O collection	 - $enable_ldms
+	   LDMS 	 runtime module  - $enable_ldms_mod
+			 collection
            Memory alignment in bytes     - $with_mem_align
            Log file env variables        - $__log_path_by_env
            Location of Darshan log files - $__log_path
